+++ CHANGES.rst
@@ -8,6 +8,8 @@ Unreleased
 - Dependencies: Fixed deprecation warnings on function signatures
   ``PDOCrateDB::prepare``, ``PDOCrateDB::lastInsertId``, ``PDOCrateDB::quote``,
   and ``PDOStatementImplementationPhp8::query``.
+- Quoting: Started supporting ``PDOCrateDB::quote(..., \PDO::PARAM_STR)``
+  when prepared statements can't be used.
 
 2025/11/13 2.2.3
 ================